native fixes

+++ test.sh
@@ -1,6 +1,9 @@
-JAVA_HOME=/usr/local/java/graalvm-ce-java11/ mvn -Pgraal clean install 
+#!/bin/sh -x 
+
+JAVA_HOME=/usr/local/java/graalvm mvn -Pgraal clean install 
 #--add-exports=java.base/jdk.internal.module=ALL-UNNAMED
-USERNAME=admin
-PASSWORD=admin
-target/jslee-js --js.debug=true --js.trace=true --js.username=$USERNAME --js.password=$PASSWORD --js.url=service:jmx:remote+http://localhost:9990 src/test/js/service-activate-deactivate.js 
-cp target/jslee-js bin
+#target/jslee-js --js.debug=true --js.trace=true --js.username=admin --js.password=admin --js.url=service:jmx:remote+http://172.69.0.6:9990 src/test/js/simple.js
+
+if [ -f "target/jslee-js" ] ; then
+  cp target/jslee-js bin
+fi
